refactor(service-common): change test extensions from spec to test
